properly built flow-closeness centrality

+++ source/adjacency_matrix_builder.h
@@ -11,7 +11,7 @@ class EdgeListBuilder // TODO: Should be called edge_list builder, should accoun
 {
     private:
         bool undirected;
-        void AddEdge(std::vector<std::vector<Adjacent>>& list, const size_t start, const size_t destination, const double weight);
+        void AddEdge(std::vector<std::vector<Adjacent>>& list, const size_t start, const size_t destination, const unsigned int weight);
 
     public:
         EdgeListBuilder();
